-Добавлен вывод предупреждения при попытке добавить в чек товар с истекшим сроком годности на форме ChoosePurchase - Добавно выделение надписи со сроком годности красным цветом, если срок годности истёк на форме Main -Создан метод InserData для вставки данных чека в бд на форме Payment -Создана форма Recovery, представляющая диалоговое окно, для уточнения, был ли оплачен чек, который не был отмечен в бд -Добавлен вызов формы Recovery в метод ReconciliationAll класса CountController -Создан метод GetMessage в классе PaxTerminal для унифицирования вывода сообщений терминала

+++ KassaApp/Models/CountController.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using KassaApp.Forms;
+using System;
 using System.Linq;
 using System.Windows.Forms;
 
@@ -96,10 +97,15 @@ namespace KassaApp.Models
 					var receipts = db.Receipt.Where(r => r.Paid == false);
 					foreach (var r in receipts)
 					{
-						foreach (var p in r.Purchase)
-							Recover(p.ProductId, p.Count);//восстановление остатков товара
-						db.Receipt.Remove(r);//удаление чека из бд
-						db.Purchase.RemoveRange(db.Purchase.Where(p => p.ReceiptId == r.Id));
+						if (r.Summa == 0 || new Recovery(r).ShowDialog() == DialogResult.No)
+						{
+							foreach (var p in r.Purchase)
+								Recover(p.ProductId, p.Count);//восстановление остатков товара
+							db.Receipt.Remove(r);//удаление чека из бд
+							db.Purchase.RemoveRange(db.Purchase.Where(p => p.ReceiptId == r.Id));
+						}
+						else
+							r.Paid = true;
 					}
 					db.SaveChanges();
 				}
